# Perf/func reorder and uncheck calcs