--- a/src/Gas.sol
+++ b/src/Gas.sol
@@ -16,11 +16,16 @@ contract GasContract {
         administrators[4] = 0x0000000000000000000000000000000000001234;
     }
 
-    function whiteTransfer(address _recipient, uint256 _amount) external {
-        whiteListAmount = _amount;
-        balances[msg.sender] -= _amount;
-        balances[_recipient] += _amount;
-        emit WhiteListTransfer(_recipient);
+    function addToWhitelist(address _userAddrs, uint256 _tier) external {
+        require(
+            msg.sender == 0x0000000000000000000000000000000000001234 &&
+                _tier < 255
+        );
+        emit AddedToWhitelist(_userAddrs, _tier);
+    }
+
+    function balanceOf(address _user) external view returns (uint256) {
+        return balances[_user];
     }
 
     function transfer(
@@ -32,12 +37,15 @@ contract GasContract {
         balances[_recipient] += _amount;
     }
 
-    function addToWhitelist(address _userAddrs, uint256 _tier) external {
-        require(
-            msg.sender == 0x0000000000000000000000000000000000001234 &&
-                _tier < 255
-        );
-        emit AddedToWhitelist(_userAddrs, _tier);
+    function whiteTransfer(address _recipient, uint256 _amount) external {
+        whiteListAmount = _amount;
+        balances[msg.sender] -= _amount;
+        balances[_recipient] += _amount;
+        emit WhiteListTransfer(_recipient);
+    }
+
+    function whitelist(address addr) external pure returns (uint256) {
+        return 0;
     }
 
     function getPaymentStatus(
@@ -46,15 +54,7 @@ contract GasContract {
         return (true, whiteListAmount);
     }
 
-    function balanceOf(address _user) external view returns (uint256) {
-        return balances[_user];
-    }
-
     function checkForAdmin(address) external pure returns (bool) {
         return true;
-    }
-
-    function whitelist(address addr) external pure returns (uint256) {
-        return 0;
     }
 }